backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
platformdirs==4.2.0
PyJWT==2.8.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
virtualenv==20.25.0
Werkzeug==3.0.1
